+++ verified-broadcast/tests/broadcast.test.js
@@ -24,44 +24,59 @@ const mockClient = {
   },
 };
 
+const mockIsAuthenticated = jest.fn(() => true);
+jest.mock('../assets/auth.private.js', () => {
+  return {
+    isAuthenticated: mockIsAuthenticated,
+  };
+});
+
+const mockSetupResources = jest.fn(async () => true);
+jest.mock('../assets/setup.private.js', () => {
+  return {
+    setupResourcesIfRequired: mockSetupResources,
+  };
+});
+
 const VERIFY_SERVICE_SID = 'default';
 
 const testContext = {
   VERIFY_SERVICE_SID,
-  BROADCAST_ADMIN_NUMBER: '+12223334444',
+  PASSCODE: 'test-code',
+  BROADCAST_NOTIFY_SERVICE_SID: 'placeholder',
+  TWILIO_PHONE_NUMBER: '+12223334444',
   getTwilioClient: () => mockClient,
 };
 
 describe('verified-broadcast/broadcast', () => {
   beforeAll(() => {
-    helpers.setup({});
+    const runtime = new helpers.MockRuntime();
+    runtime._addAsset('/setup.js', '../assets/setup.private.js');
+    runtime._addAsset('/auth.js', '../assets/auth.private.js');
+    helpers.setup(testContext, runtime);
   });
   afterAll(() => {
     helpers.teardown();
   });
 
-  test('checks verification before broadcasting', (done) => {
+  test('does not allow broadcasting without valid code', (done) => {
+    mockIsAuthenticated.mockReturnValueOnce(false);
+
     const callback = (err, result) => {
-      expect(err).toBeNull();
-      expect(result).toBeDefined();
-      expect(mockClient.verify.services).toHaveBeenCalledWith(
-        VERIFY_SERVICE_SID
-      );
-      const expectedContext = {
-        code: '123456',
-        to: '+12223334444',
-      };
-      expect(
-        mockVerificationCheck.verificationChecks.create
-      ).toHaveBeenCalledWith(expectedContext);
-      expect(result._body.success).toEqual(true);
+      expect(err).toEqual(null);
+      expect(result).toBeInstanceOf(Twilio.Response);
+      expect(result._statusCode).toEqual(401);
+      expect(result._body).toEqual({
+        error: 'INVALID CREDENTIALS',
+        success: false,
+      });
       done();
     };
-    const event = { code: '123456', body: 'example' };
+    const event = { body: 'example' };
     broadcastFunction(testContext, event, callback);
   });
 
-  test('sends broadcast if verification is successful', (done) => {
+  test('sends broadcast if password is valid', (done) => {
     const callback = (err, result) => {
       expect(err).toBeNull();
       expect(result).toBeDefined();
@@ -69,7 +84,7 @@ describe('verified-broadcast/broadcast', () => {
       expect(mockNotificationSubscribe.notifications.create).toHaveBeenCalled();
       done();
     };
-    const event = { code: '123456', body: 'broadcast' };
+    const event = { body: 'broadcast' };
     broadcastFunction(testContext, event, callback);
   });
 
@@ -84,7 +99,7 @@ describe('verified-broadcast/broadcast', () => {
       ).toHaveBeenCalledWith(expectedContext);
       done();
     };
-    const event = { code: '123456', body: 'broadcast', tag: 'pie' };
+    const event = { body: 'broadcast', tag: 'pie' };
     broadcastFunction(testContext, event, callback);
   });
 
@@ -99,41 +114,7 @@ describe('verified-broadcast/broadcast', () => {
       ).toHaveBeenCalledWith(expectedContext);
       done();
     };
-    const event = { code: '123456', body: 'broadcast' };
+    const event = { body: 'broadcast' };
     broadcastFunction(testContext, event, callback);
-  });
-
-  test('returns "incorrect token" if verification fails', (done) => {
-    const mockFailureVerificationCheck = {
-      verificationChecks: {
-        create: jest
-          .fn()
-          .mockResolvedValue({ status: 'pending', sid: 'my-new-sid' }),
-      },
-    };
-
-    const mockFailureClient = {
-      verify: {
-        services: jest.fn(() => mockFailureVerificationCheck),
-      },
-      notify: {
-        services: jest.fn(() => mockNotificationSubscribe),
-      },
-    };
-
-    const failureContext = {
-      VERIFY_SERVICE_SID,
-      getTwilioClient: () => mockFailureClient,
-    };
-
-    const callback = (err, result) => {
-      expect(err).toBeNull();
-      expect(result).toBeDefined();
-      expect(result._body.success).toEqual(false);
-      expect(result._body.error).toEqual('Incorrect token.');
-      done();
-    };
-    const event = { to: '+12223334444', body: 'message' };
-    broadcastFunction(failureContext, event, callback);
   });
 });