+++ verified-broadcast/changelog.md
@@ -2,6 +2,11 @@
 
 ## [Unreleased]
 
+## [1.1.0]
+### Added
+- Auto provisioning of resources if they are not defined
+- Moving from Verify for Admin authorization to passcode based authorization with HTTP Basic Auth
+
 ## [1.0.0]
 ### Added
 - Initial release.